#3 upgraded select
Убрал из отдельных файлов группы и итемы. С хуком ничего в голову не приходит. Понимание как писать хук есть, но как в данном случае применить не понимаю :cry:
