--- #3/src/components/customField/Items.js
@@ -1,12 +1,0 @@
-const Items = ({ name, items, handleClick }) => {
-	return items?.map((option, index) => (
-		<li
-			onClick={e => handleClick(option.value)}
-			key={name + '_option_' + index}
-		>
-			{option.text}
-		</li>
-	))
-}
-
-export default Items
